Added an example of Inversion of Control
A contrived example of a Message Writer which serializes messages and
writes them to a persistent storage. The storage is injected into the
Message Writer, allowing us to swap implementations according to
configuration using Dependency Injection and test our Message Writer
using mock objects instead of relying on concrete implementations.